requirements.txt
tensorflow==2.18.0
fastapi==0.115.0
uvicorn==0.32.0
pillow==11.0.0
numpy==1.26.4
python-multipart==0.0.12